% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cspline.R
\name{adjust_curve}
\alias{adjust_curve}
\title{Interactive curve adjustment}
\usage{
adjust_curve(
  x,
  y = NULL,
  ...,
  plotit = TRUE,
  curve = open_curve,
  ccolour = "red",
  pcolour = "navy"
)
}
\arguments{
\item{x, y}{Any means of specifying points in the plane, as accepted by xy.coords()}

\item{...}{currently ignored}

\item{plotit}{logical: should the curve be plotted (TRUE) or can it be assumed
the points are already on the graphic (FALSE)?}

\item{curve}{One of the curve type functions of this package}

\item{ccolour}{character string: colour for the curve in the plot}

\item{pcolour}{character string: colour for the curve in the plot}
}
\value{
The adjusted points which define the adjusted curve
}
\description{
A simple interactive device for adjusting a curve. Given a set of points,
the curve is plotted and may then be adjusted interactively by clicking on
any of the points, one at a time, and clicking again in the plot for its
new position.
}
