% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cspline.R
\name{as_polygon}
\alias{as_polygon}
\alias{as_polygon.default}
\alias{as_polygon.curve}
\alias{as_points}
\title{Make a Simple Polygon or Points}
\usage{
as_polygon(x, y = NULL, ...)

\method{as_polygon}{default}(x, y = NULL, ...)

\method{as_polygon}{curve}(x, y = NULL, ...)

as_points(x, y = NULL)
}
\arguments{
\item{x, y}{any specification of 2-d points, or a "curve" object}

\item{...}{additional arguments not currently used}
}
\value{
a data frame with components x and y
}
\description{
A simple polygon is here defined as a data frame with numeric components x and y
without any duplicate rows.  The order of rows is significant in defining the
associated figure.
}
\details{
A 'points' object is defined as a data frame with numeric columns x and y.
}
