% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cspline.R
\name{complexReplacement}
\alias{complexReplacement}
\alias{Re<-}
\alias{Im<-}
\alias{Mod<-}
\alias{Arg<-}
\title{Complex vector property replacement functions}
\usage{
Re(x) <- value

Im(x) <- value

Mod(x) <- value

Arg(x) <- value
}
\arguments{
\item{x}{a complex vector to be altered}

\item{value}{the numerical value vector to be used in the alteration}
}
\value{
An appropriately modified complex vector
}
\description{
Complex vector property replacement functions
}
